Fix handling of qualified names for packages with multiple dots
Currently, qualified names are generated by attempting to strip the rpc name
from the rest of the `proto-qualified-name` by using `subseq` on the `.`
character. This breaks for packages with multiple `.`s, as only the first
portion of the package name is kept. The fix simply adds the `from-end t`
keyword to the `position` call to get the last occurrence of `.` instead of the
first.

PiperOrigin-RevId: 420314977