Unit Testing for server.lisp
This changelist validates the utility of the functions defined in server.lips. More specifically, it tests the following functions: start-call-on-server, receive-message-from-client, send-initial-metadata, server-send-message, server-send-status and server-recv-close. In order to test these functions, we had to remove the in-lining of some of the "cffi" helper functions defined in server.lisp, client.lisp and wrappers.lisp since it is not possible to mock inline functions.

PiperOrigin-RevId: 463151728