Integration Test for LISP gRPC
This CL performs the integration test between the LISP gRPC client and the LISP gRPC server. We perform this test by ensuring that the message sent by the client is accurately received on the server side, and the message sent by the server is accurately being received on the client side. We make one small crucial change when we return the gRPC slice as a vector of bytes to ignore the null pointer.

PiperOrigin-RevId: 465380297